requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
fastapi==0.66.1
h11==0.12.0
idna==3.2
mysqlclient==2.0.3
peewee==3.14.4
pycodestyle==2.7.0
pydantic==1.8.2
PyJWT==2.1.0
python-multipart==0.0.5
requests==2.26.0
six==1.16.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
